Implemented logic for the adding and deletion of products from the cart in the cart service
