+++ src/app/services/cart.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { CartServiceService } from './cart-service.service';
+import { CartServiceService } from './cart.service';
 
 describe('CartServiceService', () => {
   let service: CartServiceService;
